ABC144-D

+++ ABC144/D/Program.cs
@@ -6,9 +6,9 @@ namespace D
     {
         static void Main(string[] args)
         {
-            var abx = Array.ConvertAll(Console.ReadLine().Split(' '), int.Parse);
+            var abx = Array.ConvertAll(Console.ReadLine().Split(' '), double.Parse);
             var a = abx[0]; var b = abx[1]; var x = abx[2];
-            double result = Math.Atan(2 * (double)(a * a * b - x) / (double)(a * a * a)) * 180 / Math.PI;
+            double result = Math.Atan(2 * (a * a * b - x) / (a * a * a)) * 180 / Math.PI;
             Console.WriteLine(result);
         }
     }
